Consolidate transformers for common CRD fields, registry and config
Fixes #26

Also factored the common test utilities out of images.go and into a
common/testing package.

All tests should remain intact -- eventing had a few more than serving
-- now all in a single place.
