+++ pkg/reconciler/common/config_maps.go
@@ -20,15 +20,14 @@ import (
 	mf "github.com/manifestival/manifestival"
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	eventingv1alpha1 "knative.dev/operator/pkg/apis/operator/v1alpha1"
 )
 
 // ConfigMapTransform updates the ConfigMap with the values specified in operator CR
-func ConfigMapTransform(instance *eventingv1alpha1.KnativeEventing, log *zap.SugaredLogger) mf.Transformer {
+func ConfigMapTransform(config map[string]map[string]string, log *zap.SugaredLogger) mf.Transformer {
 	return func(u *unstructured.Unstructured) error {
 		// Let any config in instance override everything else
 		if u.GetKind() == "ConfigMap" {
-			if data, ok := instance.Spec.Config[u.GetName()[len(`config-`):]]; ok {
+			if data, ok := config[u.GetName()[len(`config-`):]]; ok {
 				UpdateConfigMap(u, data, log)
 			}
 		}
