+++ pkg/reconciler/common/images_test.go
@@ -16,34 +16,38 @@ limitations under the License.
 package common
 
 import (
-	"reflect"
 	"testing"
 
+	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
-	eventingv1alpha1 "knative.dev/operator/pkg/apis/operator/v1alpha1"
+	caching "knative.dev/caching/pkg/apis/caching/v1alpha1"
+	"knative.dev/operator/pkg/apis/operator/v1alpha1"
+	util "knative.dev/operator/pkg/reconciler/common/testing"
 
 	appsv1 "k8s.io/api/apps/v1"
 )
 
-type updateDeploymentImageTest struct {
+var log = zap.NewNop().Sugar()
+
+type updateImageTest struct {
 	name       string
 	containers []corev1.Container
-	registry   eventingv1alpha1.Registry
+	registry   v1alpha1.Registry
 	expected   []corev1.Container
 }
 
-var updateDeploymentImageTests = []updateDeploymentImageTest{
+var updateImageTests = []updateImageTest{
 	{
 		name: "UsesNameFromDefault",
 		containers: []corev1.Container{{
 			Name:  "queue",
-			Image: "gcr.io/knative-releases/github.com/knative/eventing/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
+			Image: "gcr.io/knative-releases/github.com/knative/serving/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Default: "new-registry.io/test/path/${NAME}:new-tag",
 		},
 		expected: []corev1.Container{{
@@ -63,7 +67,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 				Image: "gcr.io/cmd/queue:test",
 			},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"container1": "new-registry.io/test/path/new-container-1:new-tag",
 				"container2": "new-registry.io/test/path/new-container-2:new-tag",
@@ -84,9 +88,9 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 		name: "UsesOverrideFromDefault",
 		containers: []corev1.Container{{
 			Name:  "queue",
-			Image: "gcr.io/knative-releases/github.com/knative/eventing/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
+			Image: "gcr.io/knative-releases/github.com/knative/serving/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Default: "new-registry.io/test/path/${NAME}:new-tag",
 			Override: map[string]string{
 				"queue": "new-registry.io/test/path/new-value:new-override-tag",
@@ -103,7 +107,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 			Name:  "image",
 			Image: "docker.io/name/image:tag2"},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"Unused": "new-registry.io/test/path",
 			},
@@ -119,7 +123,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 			Name:  "queue",
 			Image: "gcr.io/knative-releases/github.com/knative/eventing/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
 		},
-		registry: eventingv1alpha1.Registry{},
+		registry: v1alpha1.Registry{},
 		expected: []corev1.Container{{
 			Name:  "queue",
 			Image: "gcr.io/knative-releases/github.com/knative/eventing/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45"},
@@ -130,7 +134,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 		containers: []corev1.Container{{
 			Env: []corev1.EnvVar{{Name: "SOME_IMAGE", Value: "gcr.io/foo/bar"}},
 		}},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"SOME_IMAGE": "docker.io/my/overridden-image",
 			},
@@ -144,7 +148,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 		containers: []corev1.Container{{
 			Env: []corev1.EnvVar{{Name: "SOME_IMAGE", Value: "gcr.io/foo/bar"}},
 		}},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"OTHER_IMAGE": "docker.io/my/overridden-image",
 			},
@@ -160,7 +164,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 			Image: "gcr.io/knative-releases/github.com/knative/eventing/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45",
 			Env:   []corev1.EnvVar{{Name: "SOME_IMAGE", Value: "gcr.io/foo/bar"}},
 		}},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"queue":      "new-registry.io/test/path/new-value:new-override-tag",
 				"SOME_IMAGE": "docker.io/my/overridden-image",
@@ -184,7 +188,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 				Image: "gcr.io/cmd/queue:test",
 			},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"container1": "new-registry.io/test/path/new-container-1:new-tag",
 				"container2": "new-registry.io/test/path/new-container-2:new-tag",
@@ -215,7 +219,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 				Image: "gcr.io/cmd/queue:test",
 			},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"container1": "new-registry.io/test/path/new-container-1:new-tag",
 				"container2": "new-registry.io/test/path/new-container-2:new-tag",
@@ -245,7 +249,7 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 				Image: "gcr.io/cmd/queue:test",
 			},
 		},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			Override: map[string]string{
 				"OverrideWithDeploymentName/container1": "new-registry.io/test/path/OverrideWithDeploymentName/container-1:new-tag",
 				"OverrideWithDeploymentName/container2": "new-registry.io/test/path/OverrideWithDeploymentName/container-2:new-tag",
@@ -264,42 +268,51 @@ var updateDeploymentImageTests = []updateDeploymentImageTest{
 	},
 }
 
-func TestDeploymentTransform(t *testing.T) {
-	for _, tt := range updateDeploymentImageTests {
+func TestResourceTransform(t *testing.T) {
+	for _, tt := range updateImageTests {
 		t.Run(tt.name, func(t *testing.T) {
-			runDeploymentTransformTest(t, &tt)
+			runResourceTransformTest(t, &tt)
 		})
 	}
 }
 
-func runDeploymentTransformTest(t *testing.T, tt *updateDeploymentImageTest) {
-	unstructuredDeployment := makeUnstructured(t, makeDeployment(t, tt.name, corev1.PodSpec{Containers: tt.containers}))
-	instance := &eventingv1alpha1.KnativeEventing{
-		Spec: eventingv1alpha1.KnativeEventingSpec{
-			Registry: tt.registry,
-		},
-	}
-	deploymentTransform := DeploymentTransform(instance, log)
+func runResourceTransformTest(t *testing.T, tt *updateImageTest) {
+	// test for deployment
+	unstructuredDeployment := util.MakeUnstructured(t, util.MakeDeployment(tt.name, corev1.PodSpec{Containers: tt.containers}))
+	deploymentTransform := ImageTransform(&tt.registry, log)
 	deploymentTransform(&unstructuredDeployment)
 	validateUnstructedDeploymentChanged(t, tt, &unstructuredDeployment)
+
+	// test for daemonSet
+	unstructuredDaemonSet := util.MakeUnstructured(t, makeDaemonSet(tt.name, corev1.PodSpec{Containers: tt.containers}))
+	daemonSetTransform := ImageTransform(&tt.registry, log)
+	daemonSetTransform(&unstructuredDaemonSet)
+	validateUnstructedDaemonSetChanged(t, tt, &unstructuredDaemonSet)
 }
 
-func validateUnstructedDeploymentChanged(t *testing.T, tt *updateDeploymentImageTest, u *unstructured.Unstructured) {
+func validateUnstructedDeploymentChanged(t *testing.T, tt *updateImageTest, u *unstructured.Unstructured) {
 	var deployment = &appsv1.Deployment{}
 	err := scheme.Scheme.Convert(u, deployment, nil)
-	assertEqual(t, err, nil)
-	assertDeepEqual(t, deployment.Spec.Template.Spec.Containers, tt.expected)
+	util.AssertEqual(t, err, nil)
+	util.AssertDeepEqual(t, deployment.Spec.Template.Spec.Containers, tt.expected)
 }
 
-func makeDeployment(t *testing.T, name string, podSpec corev1.PodSpec) *appsv1.Deployment {
-	return &appsv1.Deployment{
+func validateUnstructedDaemonSetChanged(t *testing.T, tt *updateImageTest, u *unstructured.Unstructured) {
+	var daemonSet = &appsv1.DaemonSet{}
+	err := scheme.Scheme.Convert(u, daemonSet, nil)
+	util.AssertEqual(t, err, nil)
+	util.AssertDeepEqual(t, daemonSet.Spec.Template.Spec.Containers, tt.expected)
+}
+
+func makeDaemonSet(name string, podSpec corev1.PodSpec) *appsv1.DaemonSet {
+	return &appsv1.DaemonSet{
 		TypeMeta: metav1.TypeMeta{
-			Kind: "Deployment",
+			Kind: "DaemonSet",
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name: name,
 		},
-		Spec: appsv1.DeploymentSpec{
+		Spec: appsv1.DaemonSetSpec{
 			Template: corev1.PodTemplateSpec{
 				Spec: podSpec,
 			},
@@ -307,19 +320,86 @@ func makeDeployment(t *testing.T, name string, podSpec corev1.PodSpec) *appsv1.D
 	}
 }
 
-func makeUnstructured(t *testing.T, obj interface{}) unstructured.Unstructured {
-	var result = unstructured.Unstructured{}
-	err := scheme.Scheme.Convert(obj, &result, nil)
-	if err != nil {
-		t.Fatalf("Could not create unstructured object: %v, err: %v", result, err)
+type updateImageSpecTest struct {
+	name     string
+	in       string
+	registry v1alpha1.Registry
+	expected caching.ImageSpec
+}
+
+var updateImageSpecTests = []updateImageSpecTest{
+	{
+		name: "UsesNameFromDefault",
+		in:   "gcr.io/knative-releases/github.com/knative/serving/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45",
+		registry: v1alpha1.Registry{
+			Default: "new-registry.io/test/path/${NAME}:new-tag",
+		},
+		expected: caching.ImageSpec{
+			Image: "new-registry.io/test/path/UsesNameFromDefault:new-tag",
+		},
+	},
+	{
+		name: "AddsImagePullSecrets",
+		in:   "gcr.io/knative-releases/github.com/knative/serving/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45",
+		registry: v1alpha1.Registry{
+			ImagePullSecrets: []corev1.LocalObjectReference{
+				{Name: "new-secret"},
+			},
+		},
+		expected: caching.ImageSpec{
+			Image: "gcr.io/knative-releases/github.com/knative/serving/cmd/queue@sha256:1e40c99ff5977daa2d69873fff604c6d09651af1f9ff15aadf8849b3ee77ab45",
+			ImagePullSecrets: []corev1.LocalObjectReference{
+				{Name: "new-secret"},
+			},
+		},
+	},
+}
+
+func TestImageTransform(t *testing.T) {
+	for _, tt := range updateImageSpecTests {
+		t.Run(tt.name, func(t *testing.T) {
+			runImageTransformTest(t, &tt)
+		})
 	}
-	return result
+}
+func runImageTransformTest(t *testing.T, tt *updateImageSpecTest) {
+	unstructuredImage := util.MakeUnstructured(t, makeImage(t, tt))
+	instance := &v1alpha1.KnativeServing{
+		Spec: v1alpha1.KnativeServingSpec{
+			Registry: tt.registry,
+		},
+	}
+	imageTransform := ImageTransform(&instance.Spec.Registry, log)
+	imageTransform(&unstructuredImage)
+	validateUnstructedImageChanged(t, tt, &unstructuredImage)
+}
+
+func validateUnstructedImageChanged(t *testing.T, tt *updateImageSpecTest, u *unstructured.Unstructured) {
+	var image = &caching.Image{}
+	err := scheme.Scheme.Convert(u, image, nil)
+	util.AssertEqual(t, err, nil)
+	util.AssertDeepEqual(t, image.Spec, tt.expected)
+}
+
+func makeImage(t *testing.T, tt *updateImageSpecTest) *caching.Image {
+	return &caching.Image{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "caching.internal.knative.dev/v1alpha1",
+			Kind:       "Image",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: tt.name,
+		},
+		Spec: caching.ImageSpec{
+			Image: tt.in,
+		},
+	}
 }
 
 type addImagePullSecretsTest struct {
 	name            string
 	existingSecrets []corev1.LocalObjectReference
-	registry        eventingv1alpha1.Registry
+	registry        v1alpha1.Registry
 	expectedSecrets []corev1.LocalObjectReference
 }
 
@@ -327,13 +407,13 @@ var addImagePullSecretsTests = []addImagePullSecretsTest{
 	{
 		name:            "LeavesSecretsEmptyByDefault",
 		existingSecrets: nil,
-		registry:        eventingv1alpha1.Registry{},
+		registry:        v1alpha1.Registry{},
 		expectedSecrets: nil,
 	},
 	{
 		name:            "AddsImagePullSecrets",
 		existingSecrets: nil,
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			ImagePullSecrets: []corev1.LocalObjectReference{{Name: "new-secret"}},
 		},
 		expectedSecrets: []corev1.LocalObjectReference{{Name: "new-secret"}},
@@ -341,7 +421,7 @@ var addImagePullSecretsTests = []addImagePullSecretsTest{
 	{
 		name:            "SupportsMultipleImagePullSecrets",
 		existingSecrets: nil,
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			ImagePullSecrets: []corev1.LocalObjectReference{
 				{Name: "new-secret-1"},
 				{Name: "new-secret-2"},
@@ -355,7 +435,7 @@ var addImagePullSecretsTests = []addImagePullSecretsTest{
 	{
 		name:            "MergesAdditionalSecretsWithAnyPreexisting",
 		existingSecrets: []corev1.LocalObjectReference{{Name: "existing-secret"}},
-		registry: eventingv1alpha1.Registry{
+		registry: v1alpha1.Registry{
 			ImagePullSecrets: []corev1.LocalObjectReference{
 				{Name: "new-secret"},
 			},
@@ -376,32 +456,23 @@ func TestImagePullSecrets(t *testing.T) {
 }
 
 func runImagePullSecretsTest(t *testing.T, tt *addImagePullSecretsTest) {
-	unstructuredDeployment := makeUnstructured(t, makeDeployment(t, tt.name, corev1.PodSpec{ImagePullSecrets: tt.existingSecrets}))
-	instance := &eventingv1alpha1.KnativeEventing{
-		Spec: eventingv1alpha1.KnativeEventingSpec{
-			Registry: tt.registry,
-		},
-	}
-	deploymentTransform := DeploymentTransform(instance, log)
+	unstructuredDeployment := util.MakeUnstructured(t, util.MakeDeployment(tt.name, corev1.PodSpec{ImagePullSecrets: tt.existingSecrets}))
+	deploymentTransform := ImageTransform(&tt.registry, log)
 	deploymentTransform(&unstructuredDeployment)
 
 	var deployment = &appsv1.Deployment{}
 	err := scheme.Scheme.Convert(&unstructuredDeployment, deployment, nil)
 
-	assertEqual(t, err, nil)
-	assertDeepEqual(t, deployment.Spec.Template.Spec.ImagePullSecrets, tt.expectedSecrets)
-}
+	util.AssertEqual(t, err, nil)
+	util.AssertDeepEqual(t, deployment.Spec.Template.Spec.ImagePullSecrets, tt.expectedSecrets)
 
-func assertEqual(t *testing.T, actual, expected interface{}) {
-	if actual == expected {
-		return
-	}
-	t.Fatalf("expected does not equal actual. \nExpected: %v\nActual: %v", expected, actual)
-}
+	unstructuredDaemonSet := util.MakeUnstructured(t, makeDaemonSet(tt.name, corev1.PodSpec{ImagePullSecrets: tt.existingSecrets}))
+	daemonSetTransform := ImageTransform(&tt.registry, log)
+	daemonSetTransform(&unstructuredDaemonSet)
 
-func assertDeepEqual(t *testing.T, actual, expected interface{}) {
-	if reflect.DeepEqual(actual, expected) {
-		return
-	}
-	t.Fatalf("expected does not deep equal actual. \nExpected: %T %+v\nActual:   %T %+v", expected, expected, actual, actual)
+	var daemonSet = &appsv1.DaemonSet{}
+	err = scheme.Scheme.Convert(&unstructuredDaemonSet, daemonSet, nil)
+
+	util.AssertEqual(t, err, nil)
+	util.AssertDeepEqual(t, daemonSet.Spec.Template.Spec.ImagePullSecrets, tt.expectedSecrets)
 }
